split zipDataFrames since sparksession not available in 1.5
